requirements.txt
nautilus-trader
pandas
numpy
pyyaml
requests
rich
watchdog
